feat: updated links and readme.md
Updated the links that were deprecated, deleted a lot of them and added some more neutral ones, i keep a different set of links on my end, so these links are more public friendly while still getting the main idea across. Also updated the readme.me to clarify the current state of this repo and updated the #2 reference since it was outdated and no longer in use.